+++ examples/choropleth-overlay/app.js
@@ -18,10 +18,9 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import document from 'global/document';
+/* global document, window */
 import ReactDOM from 'react-dom';
 import React, {Component} from 'react';
-import window from 'global/window';
 
 import ChoroplethOverlayExample from './choropleth-overlay-example';