+++ examples/main/app.js
@@ -23,13 +23,13 @@ import React, {Component} from 'react';
 import window from 'global/window';
 import document from 'global/document';
 
+import TiltExample from './views/tilt-example';
 import NotInteractiveExample from './views/not-interactive-example';
 import CustomOverlayExample from './views/custom-overlay-example';
 import GeodataCreator from './views/geodata-creator-example';
 import ScatterplotOverlayExample from './views/scatterplot-overlay-example';
 import RouteOverlayExample from './views/route-overlay-example';
 import StyleDiffingExample from './views/style-diffing-example';
-import TiltExample from './views/tilt-example';
 import ClickExample from './views/click-example';
 
 export default class App extends Component {
@@ -43,7 +43,8 @@ export default class App extends Component {
     const common = {width: 400, height: 400, style: {float: 'left'}};
     return (
       <div>
-        <TiltExample {...common} width={ this.state.width - 30 }/>
+        <TiltExample {...common} width={this.state.width - 30}/>
+        <TiltExample {...common} width={this.state.width - 30} pressKeyToRotate={false}/>
         <RouteOverlayExample {...common}/>
         <ScatterplotOverlayExample {...common}/>
         <CustomOverlayExample {...common}/>